.agent/artifacts/g0_reaudit/build_source/requirements.txt
multi-llm-orchestrator[langchain,retrieval]>=0.9.0,<1.0.0
langchain>=1.0
langchain-classic>=1.0,<2.0
langchain-core>=0.1.0
langchain-community>=0.0.1
langchain-text-splitters>=0.0.1
python-telegram-bot>=21.0,<22.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
pyyaml>=6.0
pydantic>=2.0
redis>=5.0
httpx>=0.24.0
opensearch-py>=2.3.0
aiohttp>=3.9.0
python-json-logger>=2.0.0
prometheus-client>=0.19.0,<0.20.0
python-dotenv>=1.0.0

